requirements.txt
Flask==1.1.2
imutils==0.5.4
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
Pillow==8.1.0
